advanced

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,17 +7,17 @@
 int main(void)
 {
 	int i;
-	long double x = 0, y = 1, z;
+	long x = 0, y = 1, z;
 
 	for (i = 1; i <= 98; i++)
 	{
 		z = x + y;
 		if (i != 98)
 		{
-			printf("%Lf, ", z);
+			printf("%ld, ", z);
 		} else
 		{
-			printf("%Lf\n", z);
+			printf("%ld\n", z);
 		}
 		x = y;
 		y = z;
